requirements.txt
anyascii==0.2.0
appdirs==1.4.4
asgiref==3.5.0
astroid==2.4.2
auth==0.5.3
autopep8==1.5.4
awscli==1.20.64
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
botocore==1.21.64
#Brlapi==0.8.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.0.4
##chrome-gnome-shell==0.0.0
colorama==0.4.3
cryptography==3.3.1
#cupshelpers==1.0
dbus-python==1.2.16
decorator==5.0.9
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
distro-info==1.0
Django==4.0.2
django-allauth==0.44.0
django-bower==5.2.0
django-cogwheels==0.3
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-extensions==3.1.3
django-filter==2.4.0
django-modelcluster==5.2
django-taggit==2.1.0
django-treebeard==4.3.1
djangorestframework==3.11.2
dnspython==1.16.0
docutils==0.15.2
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
eventlet==0.30.0
fabric==2.5.0
Fabric3==1.14.post1
falcon==2.0.0
filelock==3.3.1
flake8==3.8.4
greenlet==0.4.17
gunicorn==20.0.4
html5lib==0.999999999
httplib2==0.18.1
idna==2.10
importlib-metadata==3.3.0
invoke==1.4.1
ipython==7.24.1
ipython_genutils==0.2.0
isort==5.6.4
jedi==0.18.0
jmespath==0.10.0
l18n==2020.6.1
lazy-object-proxy==1.4.3
#louis==3.16.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mongoengine==0.22.1
more-itertools==4.2.0
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.7
package-name==0.1
pandas==1.1.5
paramiko==2.7.2
parso==0.8.2
pathlib2==2.3.6
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==4.3.0
platformdirs==2.4.0
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.9.0
PyGObject==3.38.0
PyJWT==2.0.0
pylint==2.6.0
pymongo==3.11.2
PyNaCl==1.4.0
PyPDF2==1.26.0
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.2.1
python-dateutil==2.8.1
python-debian==0.1.39
python-debianbts==3.1.0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.27
PyYAML==5.4.1
reportbug==7.10.3+deb11u1
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
snowballstemmer==2.0.0
soupsieve==2.2.1
sqlparse==0.4.1
stevedore==3.5.0
tablib==3.0.0
telepath==0.1.1
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.10.0.0
ujson==4.0.1
unattended-upgrades==0.1
Unidecode==0.4.21
urllib3==1.26.4
virtualenv==20.4.0+ds
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
wagtail==2.16
wagtailcommonblocks==0.1
wagtailfontawesome==1.2.1
wagtailmenus==3.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
Willow==1.4
wrapt==1.12.1
xdg==5
xhtml2pdf==0.2.4
xlrd==2.0.1
XlsxWriter==1.4.0
xlwt==1.3.0
yapf==0.30.0
zipp==3.4.0
